// api/src/models/company.model.ts
export interface ICompany {
  id: string;
  accountId: string;
  name: string;
  nit: string;
  cellphone?: string;
  employeesNumber?: number;
  creationDate?: string;
  classificationId?: string;
  web?: string;
}

export type ICreateCompanyDTO = Omit<
  ICompany,
  | "id"
  | "cellphone"
  | "employeesNumber"
  | "creationDate"
  | "classificationId"
  | "web"
>;

export type IUpdateCompanyDTO = Omit<ICompany, "id">;